Before delete temporal file for big files, it is necessary to check if we are signing big files
git-svn-id: http://jira.uji.es/repos/uji/trunk@2944 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ uji-crypto-openxades/src/main/java/es/uji/security/crypto/openxades/OpenXAdESSignatureFactory.java
@@ -108,7 +108,18 @@ public class OpenXAdESSignatureFactory implements ISignFormatProvider
         sdoc.getDataFile(0).setMimeType(xadesFileMimeType);
 
         signatureResult = signDoc(sdoc, signatureOptions);
-        random.delete();
+
+        if (signatureOptions.getSwapToFile())
+        {
+        	try
+        	{
+        		random.delete();
+        	}
+        	catch (Exception e)
+        	{
+        		e.printStackTrace();
+        	}
+        }
 
         return signatureResult;
     }
